Clean up item component
